Extract Session into a separate module
Subset of https://github.com/mfussenegger/nvim-dap/pull/114
